backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-multipart==0.0.6

# AI/ML
google-generativeai==0.8.5

# Speech Processing (Optional)
azure-cognitiveservices-speech==1.34.0

# Database
aiosqlite==0.19.0

# External APIs
notion-client==2.2.1
youtube-transcript-api==0.6.2
obsws-python==1.6.1
httpx==0.26.0

# Utils
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
python-json-logger==2.0.7

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
ruff==0.1.11
black==23.12.1
mypy==1.8.0